requirements.txt
diffusers==0.21.0
accelerate
transformers
torch
torchvision

numpy
Pillow
pydantic==2.5.3